corrections des propriétés de Piece et RevetementMur
- RevetementMur : les positions sont maintenant des doubles !
- Piece : normalisation (descr. + ordre)

+++ projetbat/src/main/java/fr/insa/dorgli/projetbat/RevetementMur.java
@@ -2,12 +2,12 @@ package fr.insa.dorgli.projetbat;
 
 public class RevetementMur implements ToStringShort {
 	TypeRevetement typeRevetement;
-	int pos1L;
-	int pos1H;
-	int pos2L;
-	int pos2H;
+	double pos1L;
+	double pos1H;
+	double pos2L;
+	double pos2H;
 
-	public RevetementMur(TypeRevetement typeRevetement, int pos1L, int pos1H, int pos2L, int pos2H) {
+	public RevetementMur(TypeRevetement typeRevetement, double pos1L, double pos1H, double pos2L, double pos2H) {
 		this.typeRevetement = typeRevetement;
 		this.pos1L = pos1L;
 		this.pos1H = pos1H;
@@ -23,7 +23,7 @@ public class RevetementMur implements ToStringShort {
 		this.typeRevetement = typeRevetement;
 	}
 
-	public int getPos1L() {
+	public double getPos1L() {
 		return pos1L;
 	}
 
@@ -31,7 +31,7 @@ public class RevetementMur implements ToStringShort {
 		this.pos1L = pos1L;
 	}
 
-	public int getPos1H() {
+	public double getPos1H() {
 		return pos1H;
 	}
 
@@ -39,7 +39,7 @@ public class RevetementMur implements ToStringShort {
 		this.pos1H = pos1H;
 	}
 
-	public int getPos2L() {
+	public double getPos2L() {
 		return pos2L;
 	}
 
@@ -47,7 +47,7 @@ public class RevetementMur implements ToStringShort {
 		this.pos2L = pos2L;
 	}
 
-	public int getPos2H() {
+	public double getPos2H() {
 		return pos2H;
 	}
